moves dropdown into a folder
